+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add 'difficulty' to getSurveys API [#22](https://github.com/rokmetro/surveys-building-block-fork/issues/22)
 - Add ranking to the custom and global leaderboard [#24](https://github.com/rokmetro/surveys-building-block-fork/issues/24)
 - Add custom leaderboard [#25](https://github.com/rokmetro/surveys-building-block-fork/issues/25)
+- Custom leaderboard notifications [#29](https://github.com/rokmetro/surveys-building-block-fork/issues/29)
 
 ### Fixed
 - Update Quizzes Pipeline to Correctly Filter and Paginate [#17](https://github.com/rokmetro/surveys-building-block-fork/issues/17)